FIX don't delete given test env

+++ conftest.py
@@ -9,6 +9,7 @@ from benchopt.utils.shell_cmd import create_conda_env
 DEFAULT_GLOBAL['debug'] = True
 DEFAULT_GLOBAL['raise_install_error'] = True
 
+_clean_env = False
 _TEST_ENV_NAME = None
 
 
@@ -40,13 +41,15 @@ def pytest_collection_modifyitems(config, items):
 
 @pytest.fixture
 def test_env_name(request):
-    global _TEST_ENV_NAME
+    global _TEST_ENV_NAME, _clean_env
 
     if _TEST_ENV_NAME is None:
         env_name = request.config.getoption("--test-env")
         recreate = request.config.getoption("--recreate")
         if env_name is None:
+            _clean_env = True
             env_name = f"_benchopt_test_env_{uuid.uuid4()}"
+
         _TEST_ENV_NAME = env_name
 
         create_conda_env(_TEST_ENV_NAME, recreate=recreate)
@@ -58,5 +61,5 @@ def test_env_name(request):
 def delete_test_env():
     global _TEST_ENV_NAME
 
-    if _TEST_ENV_NAME is not None:
+    if _TEST_ENV_NAME is not None and _clean_env:
         delete_conda_env(_TEST_ENV_NAME)
